requirements.txt
anthropic==0.42.0
openai-whisper==20240930
watchdog==6.0.0
python-docx==1.1.2
flask==3.1.0
requests==2.32.3
google-api-python-client==2.190.0
google-auth-oauthlib==1.2.4
google-auth-httplib2==0.2.0
